prefactor- split preload by lab
