+++ apps/src/p5lab/P5Lab.js
@@ -1229,15 +1229,9 @@ export default class P5Lab {
    *         loading the game.
    */
   onP5Preload() {
-    Promise.all([
-      this.isBlockly
-        ? this.preloadSpriteImages_()
-        : this.preloadAnimations_(this.level.pauseAnimationsByDefault),
-      this.maybePreloadBackgrounds_(),
-      this.runPreloadEventHandler_()
-    ]).then(() => {
-      this.p5Wrapper.notifyPreloadPhaseComplete();
-    });
+    this.preloadLabAssets()
+      .then(() => this.runPreloadEventHandler_())
+      .then(() => this.p5Wrapper.notifyPreloadPhaseComplete());
     return false;
   }
 
@@ -1250,39 +1244,6 @@ export default class P5Lab {
     }
   }
 
-  // Preloads background images if this is Sprite Lab
-  maybePreloadBackgrounds_() {
-    if (!this.isBlockly) {
-      return Promise.resolve();
-    }
-    return this.p5Wrapper.preloadBackgrounds();
-  }
-
-  /**
-   * Wait for animations to be loaded into memory and ready to use, then pass
-   * those animations to P5 to be loaded into the engine as animations.
-   * @param {Boolean} pauseAnimationsByDefault whether animations should be paused
-   * @returns {Promise} which resolves once animations are in memory in the redux
-   *          store and we've started loading them into P5.
-   *          Loading to P5 is also an async process but it has its own internal
-   *          effect on the P5 preloadCount, so we don't need to track it here.
-   * @private
-   */
-  async preloadAnimations_(pauseAnimationsByDefault) {
-    await this.whenAnimationsAreReady();
-    // Animations are ready - send them to p5 to be loaded into the engine.
-    return this.p5Wrapper.preloadAnimations(
-      getStore().getState().animationList,
-      pauseAnimationsByDefault
-    );
-  }
-
-  async preloadSpriteImages_() {
-    await this.whenAnimationsAreReady();
-    return this.p5Wrapper.preloadSpriteImages(
-      getStore().getState().animationList
-    );
-  }
   /**
    * Check whether all animations in the project animation list have been loaded
    * into memory and are ready to use.
