+++ apps/src/p5lab/gamelab/GameLab.js
@@ -1,4 +1,5 @@
 import msg from '@cdo/gamelab/locale';
+import {getStore} from '@cdo/apps/redux';
 import P5Lab from '../P5Lab';
 import {P5LabType} from '../constants';
 import project from '@cdo/apps/code-studio/initApp/project';
@@ -31,6 +32,31 @@ export default class GameLab extends P5Lab {
     }
 
     return super.init(config);
+  }
+
+  /**
+   * Wait for animations to be loaded into memory and ready to use, then pass
+   * those animations to P5 to be loaded into the engine as animations.
+   * @param {Boolean} pauseAnimationsByDefault whether animations should be paused
+   * @returns {Promise} which resolves once animations are in memory in the redux
+   *          store and we've started loading them into P5.
+   *          Loading to P5 is also an async process but it has its own internal
+   *          effect on the P5 preloadCount, so we don't need to track it here.
+   * @private
+   */
+  async preloadAnimations_(pauseAnimationsByDefault) {
+    await this.whenAnimationsAreReady();
+    // Animations are ready - send them to p5 to be loaded into the engine.
+    return this.p5Wrapper.preloadAnimations(
+      getStore().getState().animationList,
+      pauseAnimationsByDefault
+    );
+  }
+
+  preloadLabAssets() {
+    return Promise.all([
+      this.preloadAnimations_(this.level.pauseAnimationsByDefault)
+    ]);
   }
 
   resetHandler(ignore) {
